Add better git submodule handling (#10)
* Refactor git submodule installation

Let git handle the installation of the submodules included in the
`.gitmodules` file, instead of maintaining a manual list in the
install script.

* Add git index update to ignore missing symlinks

After removing the symlinks, git reports them as deleted. Adding them
to `.gitignore` do not make a difference because they are in the
history of the repo. This is a workaround to prevent the annoyance.

* Add ignore for backups created by install script